Call Await twice on rejected promise can cause runtime error (#11)
If you call Await twice on a rejected promise, the second time will throw a `nil` exception, which can lead to a total application crash with a runtime error. This is not caught by the Delphi exception handler. While this is normally an indication of an error in your codeflow, this is not expected behavior.

It is now fixed by supporting multiple calls to Await on a rejected promise. Only the first time the original exception is raised, on the subsequent calls a clone of the original exception object is raised.